requirements.txt
amqp==5.0.6
appdirs==1.4.4
attrs==20.3.0
billiard==3.6.4.0
black==20.8b1
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
dataclasses==0.8
#datasync @ file:///home/chitrankd/Documents/assignments/postman/large_file_processor/.tox/.tmp/package/1/datasync-0.1.0.zip
flake8==3.9.0
greenlet==1.0.0
importlib-metadata==3.10.0
iniconfig==1.1.1
kombu==5.0.2
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.19.5
packaging==20.9
pandas==1.1.5
pathspec==0.8.1
pluggy==0.13.1
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.3
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
six==1.15.0
SQLAlchemy==1.4.5
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.1
